Speed optimized <txp:else />.
Some numbers for various contents of <txp:if*> ... </txp:if*>:
* 2-5x faster: true <txp:tag1 /> <txp:else /> false <txp:tag2 />
* 4x faster: text
* 6x faster: true <txp:else /> false
* 6x faster: <txp:tag />
* 8x faster: true <txp:else /> false <txp:tag />
* 12x faster: <txp:tag1 /><txp:tag2 /><txp:tag3 />

I've added a specific optimisation for if-constructs that don't contain a <txp:else /> tag or contain no TXP tags before the <txp:else /> tag. This why all the examples given above (except the first one) are so faster than the more generic first examples. The last example shows the implementation works very well (even better!) for larger sets of tags contained in the IF-construct.

The speed increase in the generic first example depends on where the <txp:else /> tag is placed. If it's at the beginning, you'll get a 5x speedup. If it's at the end, it's still at least 2x faster than before.