In Textile 2.5 'Lite' mode no longer does what it used to.
In 2.5 it does actual parsing in more the intended limited
mode similarly to Restricted+Lite. In 2.4, it only ran few of 
the supporting functions, never invoking the parser.

Since now it does parsing, doing this would add HTML to
the titles.

As of Textpattern 4.5.0 section titles are returned through
HTML encoder. As such any encoding during saving is
unnecessary and added extra entities would get presented
as a plain text during display.