eLink() properly encodes values used in the query string.
This doesn't yet touch the bizarre $linktext sanitization,
which can render this function close to useless outside core
use cases.

The function renders entities as HTML instead of
presenting them as text, while encodes brackets and
quotes. This makes encoding the link labels impossible.

This behavior has its roots in the article title and the way
it's stored in the database.