Introduces centralized pane handler.
One PHP path handles all pane visibility state sync events.
The requests are validated using tokens.

Due to inconsistent ID naming, had to add extra data
attribute for hosting the panel identifier. If all IDs were
prefixed, this wouldn't be needed. Due to plugins the
event can not be prepended to the name automatically
either due to it causing some oddly named '$pane's.

May later on automate the pane generation so that it's
never defined. Instead we could calculate it from the $event
and the $anchor_id.