Fix form and contained statement checks.
Number of tags would ignore 'empty' contained statements.
As longs as it's set, it's set and should be used. Even an
empty string or a zero is a valid statement and shouldn't
be silently ignored.

Same goes for forms. Eventually we will have support for
zeros as the name for any content, including categories,
links, authors, pages and forms.