Introduces tag registering.
Textpattern_Tag_Registry class can be used to sign in
tags to the registry. This allows tags use any callbacks,
offers security due to tags not being mapped directly to
functions and allows automation.

Currently any old function-to-tag mappings work as a
fallback, but will issue a notice in debugging mode.